option for a member to leave the project

+++ front-end/src/components/app/Projects/ProjectDetails/TopPanel.tsx
@@ -4,11 +4,11 @@ import { NavLink, useNavigate } from "react-router-dom";
 import { ProjectType } from "../../../../utils/zodSchemas";
 import { statusColor } from "../../../../utils/data";
 import { useDispatch, useSelector } from "react-redux";
-import clsx from "clsx";
 import { AppDispatch, RootState } from "../../../../state/store";
 import DeleteAlert from "../../../DeleteAlert";
 import { useState } from "react";
 import { deleteProject } from "../../../../state/projects/projectThunk";
+import { removeMember } from "../../../../state/projects/membersThunk";
 
 type TopPanelProps = {
   id: ProjectType["_id"];
@@ -27,7 +27,8 @@ export default function TopPanel({ id, status, owner }: TopPanelProps) {
   const [isOpen, setIsOpen] = useState(false);
 
   const handleDelete = async () => {
-    dispatch(deleteProject({ id }));
+    if (projectOwner) dispatch(deleteProject({ id }));
+    else dispatch(removeMember({ id, memberId: user!._id }));
     navigate("/projects", { replace: true });
     location.reload();
   };
@@ -35,12 +36,7 @@ export default function TopPanel({ id, status, owner }: TopPanelProps) {
   const projectOwner = user?._id === owner._id;
 
   return (
-    <div
-      className={clsx(
-        "flex items-center text-lg bg-slate-900 px-4 py-2 rounded-lg shadow-[0_0_6px_1px_#314158]",
-        projectOwner ? "justify-between" : "justify-center"
-      )}
-    >
+    <div className="flex items-center text-lg bg-slate-900 px-4 py-2 rounded-lg shadow-[0_0_6px_1px_#314158] justify-between">
       {/* STATUS */}
       <div className="flex items-center gap-2">
         <div className={`${statusColor[status]} w-4 h-4 rounded-full`} />
@@ -48,8 +44,8 @@ export default function TopPanel({ id, status, owner }: TopPanelProps) {
       </div>
 
       {/* BUTTONS */}
-      {projectOwner && (
-        <div className="flex gap-5">
+      <div className="flex gap-5">
+        {projectOwner && (
           <NavLink
             to={`/projects/${id}/update`}
             className="flex items-center gap-2 text-primary cursor-pointer hover:text-violet-300 transition-colors active:text-violet-300"
@@ -57,21 +53,21 @@ export default function TopPanel({ id, status, owner }: TopPanelProps) {
             <span>Edit</span>
             <FontAwesomeIcon icon={faEdit} />
           </NavLink>
-          <button
-            className="flex items-center gap-2 text-red-400 cursor-pointer hover:text-red-300 transition-colors active:text-red-300"
-            onClick={() => setIsOpen(true)}
-          >
-            <span>Delete</span>
-            <FontAwesomeIcon icon={faTrash} />
-          </button>
-        </div>
-      )}
+        )}
+        <button
+          className="flex items-center gap-2 text-red-400 cursor-pointer hover:text-red-300 transition-colors active:text-red-300"
+          onClick={() => setIsOpen(true)}
+        >
+          <span>{projectOwner ? "Delete" : "Leave"}</span>
+          <FontAwesomeIcon icon={faTrash} />
+        </button>
+      </div>
 
       <DeleteAlert
         isOpen={isOpen}
         setIsOpen={setIsOpen}
         handleDelete={handleDelete}
-        message="this project"
+        message={`Are you sure you want to ${projectOwner ? "delete" : "leave"} this project?`}
         loading={loadingDelete}
         error={error}
       />
